refactor(app): extract ErrorBoundary into common component
Refs: STORY-0005

Generated-by: GitHub Copilot (GPT-5.3-Codex)
