+++ docs/stories/INDEX.md
@@ -22,7 +22,7 @@ Use this file as the local agent-readable backlog. Keep IDs sequential and never
 | STORY-0002 | CLOSED | Remove JetBrains .idea folder          | unassigned | Completed 2026-03-06: removed tracked .idea files and enforced ignore rule                    |
 | STORY-0003 | CLOSED | Remove index.html import map           | unassigned | Completed 2026-03-06: removed import map, validated build/tests/dev boot                      |
 | STORY-0004 | CLOSED | Move Tailwind config out of HTML       | unassigned | Completed 2026-03-06: moved Tailwind config to repo files and fixed theme/spacing regressions |
-| STORY-0005 | OPEN   | Extract ErrorBoundary component        | unassigned | Move from App.tsx                                                                             |
+| STORY-0005 | CLOSED | Extract ErrorBoundary component        | unassigned | Completed 2026-03-06: extracted ErrorBoundary to common component and added focused tests     |
 | STORY-0006 | OPEN   | Replace custom benchmark tooltip       | unassigned | Use Floating UI or Radix                                                                      |
 | STORY-0007 | OPEN   | Enable strict typing cleanup slice     | unassigned | Remove local any usage                                                                        |
 | STORY-0008 | OPEN   | Improve service error result model     | unassigned | Propagate actionable errors                                                                   |
